refactored out ParamToUnitString
